+++ requirements.txt
@@ -2,3 +2,4 @@ qiskit
 matplotlib
 torch
 torchvision
+-e ./qiskit-textbook-src
